Refactor to have setup plugin command handler
